Split side panel into container + body; float with 6px gap
- GlobalSidePanel is now a thin container shell (rounded-2xl, border)
- SidePanelBody holds all the existing content + modals
- GlobalLayout wraps the aside in 6px padding on left/top/bottom so
  the panel floats; total footprint stays 320px (314 + 6)
